Extract shared POSBookingSummaryView for subtitle, badges, and date formatting
